Use the activiti page instead of task page in the demo shell
