+++ demo-shell-ng2/app/components/activiti/activiti-demo.component.css
@@ -2,3 +2,10 @@
     cursor: pointer;
     font-weight: bold;
 }
+
+.task-column {
+    background-color: #f5f5f5;
+    padding: 10px 10px 10px 10px;
+    border-left: solid 2px rgb(31,188,210);
+    border-right : solid 2px rgb(31,188,210);
+}
