+++ demo-shell-ng2/app/components/activiti/activiti-demo.component.html
@@ -1,22 +1,30 @@
 <div class="mdl-grid">
-
-    <div class="mdl-cell mdl-cell--2-col">
-        <div *ngIf="!hasTasks()">No tasks found</div>
-        <div *ngIf="hasTasks()">
-            <ul class="mdl-list">
-                <li *ngFor="let task of tasks"
-                    class="mdl-list__item activiti-task-list__item"
-                    (click)="onTaskClicked(task, $event)">
-                    <span class="mdl-list__item-primary-content">
-                    <i class="material-icons mdl-list__item-icon">launch</i>
-                    {{task.name || 'Nameless task'}}
-                </span>
-            </ul>
-        </div>
+    <div class="mdl-cell mdl-cell--2-col task-column"
+         style="">
+        <ul class="demo-list-item mdl-list">
+            <li class="mdl-list__item">
+            <span class="mdl-list__item-primary-content">
+              <label class="mdl-radio mdl-js-radio mdl-js-ripple-effect" for="option-1">
+                  <input type="radio" value="task-list"
+                         (change)="setChoice($event)" name="options" id="option-1" checked class="mdl-radio__button">
+                  <span class="mdl-radio__label">Task List</span>
+              </label>
+            </span>
+            </li>
+            <li class="mdl-list__item">
+                <label class="mdl-radio mdl-js-radio mdl-js-ripple-effect" for="option-2">
+                    <input type="radio" value="process-list"
+                           (change)="setChoice($event)" name="options" id="option-2" class="mdl-radio__button">
+                    <span class="mdl-radio__label">Process List</span>
+                </label>
+            </li>
+        </ul>
     </div>
-
-    <div class="mdl-cell mdl-cell--10-col">
-        <activiti-form [taskId]="selectedTask?.id"></activiti-form>
+    <div class="mdl-cell mdl-cell--3-col task-column"
+         style="">
+        <activiti-tasklist *ngIf="isTaskListSelected()" [data]="data" (rowClick)="onRowClick($event)"></activiti-tasklist>
     </div>
-
+    <div class="mdl-cell mdl-cell--7-col task-column">
+        <activiti-task-details [taskId]="currentTaskId"></activiti-task-details>
+    </div>
 </div>
